+++ eras/conway/impl/src/Cardano/Ledger/Conway/Governance.hs
@@ -38,6 +38,10 @@ module Cardano.Ledger.Conway.Governance (
   AnchorDataHash,
   ensConstitutionL,
   rsEnactStateL,
+  Constitution (..),
+  ConstitutionData (..),
+  constitutionHashL,
+  constitutionScriptL,
 ) where
 
 import Cardano.Ledger.BaseTypes (EpochNo (..), ProtVer (..), StrictMaybe)
@@ -73,11 +77,9 @@ import Cardano.Ledger.Conway.Governance.Procedures (
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.Keys (KeyHash (..), KeyRole (..))
-import Cardano.Ledger.SafeHash (SafeHash)
 import Cardano.Ledger.Shelley.Governance
 import Control.DeepSeq (NFData)
 import Data.Aeson (KeyValue, ToJSON (..), object, pairs, (.=))
-import Data.ByteString (ByteString)
 import Data.Default.Class (Default (..))
 import Data.Map.Strict (Map)
 import Data.Sequence.Strict (StrictSeq)
@@ -174,7 +176,7 @@ instance EraPParams era => FromCBOR (ConwayGovState era) where
 data EnactState era = EnactState
   { ensCommittee :: !(StrictMaybe (Set (KeyHash 'Voting (EraCrypto era)), Rational))
   -- ^ Constitutional Committee
-  , ensConstitution :: !(SafeHash (EraCrypto era) ByteString)
+  , ensConstitution :: !(Constitution era)
   -- ^ Hash of the Constitution
   , ensProtVer :: !ProtVer
   , ensPParams :: !(PParams era)
@@ -183,7 +185,7 @@ data EnactState era = EnactState
   }
   deriving (Generic)
 
-ensConstitutionL :: Lens' (EnactState era) (SafeHash (EraCrypto era) ByteString)
+ensConstitutionL :: Lens' (EnactState era) (Constitution era)
 ensConstitutionL = lens ensConstitution (\x y -> x {ensConstitution = y})
 
 instance EraPParams era => ToJSON (EnactState era) where
@@ -344,4 +346,4 @@ toConwayGovernancePairs cg@(ConwayGovernance _ _) =
 
 instance EraPParams (ConwayEra c) => EraGovernance (ConwayEra c) where
   type GovernanceState (ConwayEra c) = ConwayGovernance (ConwayEra c)
-  getConstitutionHash g = Just $ g ^. cgRatifyL . rsEnactStateL . ensConstitutionL
+  getConstitutionHash g = Just $ g ^. cgRatifyL . rsEnactStateL . ensConstitutionL . constitutionHashL