+++ libs/cardano-ledger-core/src/Cardano/Ledger/Address.hs
@@ -106,6 +106,7 @@ import qualified Data.ByteString.Base16 as B16
 import qualified Data.ByteString.Lazy as BSL
 import Data.ByteString.Short as SBS (ShortByteString, fromShort, index, length, toShort)
 import qualified Data.ByteString.Unsafe as BS (unsafeDrop, unsafeIndex, unsafeTake)
+import Data.Default.Class (Default (..))
 import Data.Function (fix)
 import Data.Map.Strict (Map)
 import Data.Maybe (fromMaybe)
@@ -175,6 +176,9 @@ data RewardAcnt c = RewardAcnt
   , getRwdCred :: !(Credential 'Staking c)
   }
   deriving (Show, Eq, Generic, Ord, NFData, ToJSONKey, FromJSONKey)
+
+instance Crypto c => Default (RewardAcnt c) where
+  def = RewardAcnt def def
 
 instance Crypto c => ToJSON (RewardAcnt c) where
   toJSON ra =